Connected separate sections, park fix.
Fixed a bug in the code for placing a park (incorrect conditional).

Now, after removing random streets, we then go through and connect up
all the discrete sections, and make sure that all areas are available.

Also added a new initial buildCity function to set the number of parks
without having to also specify the sizes.