Beginning of steering forces.
Added two base scripts for steering. SteeringForce is an interface
that all steering behaviors must implement. SteeringObject is a
script to be placed on any object that we want steering behaviors
applied to. Also added a small scene to do testing in. It will be
removed once steering is finalized.