Changes to how victims path and path generation.
Removed the object's position from the front of the path.

Victims now have a destination field. This holds the current destination
and it is shared between victims that are also sharing paths. Also,
a lot of changes checks are done before assuming a new path. Victims
advertise who they got their path from (i.e., who initiated the path),
and if that person is too far way, then the path is not taken. If the
object we are considering cannot be seen then the path is not taken.
Finally, if we determine that the path is good (originator is not too
far away and object we are looking at is in sight) then we check if we
can see the first destination. If that is not the case, we push that
first destintation into the path, and set the destination as the
position of the object we just got the path from.