More genetic fun.
First, added warning ignore for no GetHashCode() in PathFinder because
it was bothering me.

Due to a little out of order happenings when running the genetic
algorithm, I added a clone values function to VictimData, which takes
another VictimData class and copies all the values from this other
object into itself.

I believe that I have finished the main part of the algorithm, except
there is an issue early on, which I haven't been able to resolve.
(Line 245, null argument to be specific)