Genetic is pretty much done.
Need to add mutations.

It definitely needs to be tested more than my limited testing
and won't really show much until we add in the ability to do damage
(by which I mean I need to figure out how it's being done and track it)
because for now the fitness is pretty boring, just time alive.