Fixed up population generation.
Added code to handle the event when there is only a single victim.
This will never happen in game, but it can happen when we're testing.
It just remakes the same victim and continues on.

Also, changed structure of code so that we no longer create a version
of VictimData using 'new'. Now, victims are created in chunks. First,
the new population are created, instantiating one, then filling in
its data with values from parents. After they are all done, we walk
through the old victims that are sticking around and create them.