Mutations and some minor changes.
Fixed possibility for a victim to have a single parent (weird).
Changed ToString method of GeneticVictim to fit the correct
order of listed attributes.

Added in mutations. With a 50% chance, a victim is selected to be
mutated. Then we randomly select between 1 and 3 attributes to mutate.
How an attribute is mutated depends somewhat on which one it is
(see code for details). Each one is then clamped between 0 and some
upper bound (again, depends on the attribute).