Add simple game over screen and game over triggers
I dont know how Chris worked his main menu magic, but I attempted to do
the same thing for the game over screen. Still needs a "Play" and "Quit"
button for the user. Also prompted game overs when the timer ran out and
health ran out. Its a sudden switch now, so we probably might want to
add a message box/delay so the user knows that he is about to game over